requirements.txt
Flask
psycopg2-binary
SQLAlchemy
joblib
google-generativeai
numpy
scikit-learn==1.6.1
gunicorn